requirements.txt
numpy>=1.24
pandas>=2.0
Pillow>=10.0
PyYAML>=6.0
tqdm>=4.66
matplotlib>=3.8
opencv-python-headless>=4.9
scikit-learn>=1.4
pvlib>=0.10
einops>=0.7
requests>=2.31
gdown>=5.2
huggingface_hub>=0.30
fvcore>=0.1.5.post20221221
ptflops>=0.7.3
timm>=1.0.0
causal-conv1d>=1.4.0
mamba-ssm>=2.2.4
